Tarefa2_2018li1g107

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

Funções auxiliares da Tarefa 2.

auxDisparoJog

Arguments

:: Jogador
O Jogador que efetua a jogada

-> Jogada
A Jogada a efetuar

-> Jogador
O Jogador resultante após jogada

Esta funçao auxiliar premite-nos atualizar um dado jogador apos uma jogada do tipo Disparo Arma.

(Devolve nos a informaçao desse jogador,apos um determinado disparo, no qual gastou uma muniçao.)

auxAtualizaDisp

Arguments

:: [Jogador]
A lista dos Jogadores

-> Int
O identificador do Jogador de acordo com o indice da lista

-> Jogada
A Jogada a efetuar

-> [Jogador]
A lista de Jogadores resultante após jogada

Funçao que nos premite obter a lista de jogadores atualizada tendo em conta a nova informaçao obtida na funçao anterior.

(Dada uma lista de jogadores, atualizar a informaçao obtida na auxDisparo relativamente a um dado jogador, devolvendo-nos de novo a lista, atualizada.)

auxEscolheJ

Arguments

:: [Jogador]
Lista dos Jogadores

-> Int
Identificador do Jogador de acordo com o indice da lista

-> Jogador
Jogador escolhido

Da nos um determinado jogador, da lista de jogadores, conforme a sua posiçao na mesma.

auxPosDisparo

Arguments

:: Jogador
O Jogador

-> Posicao
Posicao do Disparo

Dado um disparo, este posiciona-se na posiçao em frente do tanque que o realizou, conforme a sua direçao.

auxListaDisparo

Arguments

:: Disparo
Disparo efetuado por um jogador

-> [Disparo]
Lista anterior

-> [Disparo]
Lista resultante apos disparo

Apos uma jogada do tipo Disparo a lista correspondente aos disparos dos Jogadores é atualizada.

(É lhe acrescentada o disparo anteriormente efetuada por um Jogador)

auxListaJog

Arguments

:: [Jogador]
Lista dos jogadores

-> Int
Identificador do Jogador conforme o indice da lista

-> Jogada
Jogada a efetuar

-> Mapa
Mapa atual onde decorre o jogo

-> [Jogador]
Lista dos jogadores resultante apos jogada

Funcao que em conjunto com auxPossivelMove e, consquentemente, a auxPossivel2 e coordenadalista, no diz se um dado jogador, dada uma Jogada do tipo Movimenta, pode efetuar tal jogada, e se sim alterando a sua posiçao e atualiznado a nova informacao na lista de jogadores.

auxPossivelMove :: Jogador -> [Jogador] -> Jogada -> Mapa -> Jogador

Da nos a nova Posiao de um Jogador apos uma Jogada do tipo Movimenta valida

auxPossivel2 :: Jogador -> [Jogador] -> Bool

Dado um Jogador, verifica se existe agum outro Jogador perto da sua area,tendo em conta a direcao em que o mesmo se encontra.

coordenadalista

Arguments

:: Posicao
Posicao a verificar

-> [Jogador]
dLista de todos od jogadores

-> Bool
Verificaçao da posicao como livre ou ocupada

Verifica se , dada uma posicao, algum Jogador lá se encontra.

auxApanhaChq :: Posicao -> [Jogador] -> [Bool]

Verifica se um certo Jogador se encontra na area de choque