Tarefa3_2018li1g107

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Introduçao:

A ideia nesta fase de desenvolvimento anda em torno de tornar o jogo em si mais eficiente, ou seja, nesta tarefa, através de funções como a função comprime e descomprime, foi nos possivel poupar espaço em disco quando o estado do jogo for gravado.

Objetivo:

O objetivo desta tarefa é, dada uma descrição do estado do jogo implementar um mecanismo de compressão / descompressão que permita poupar caracteres e consequentemente espaço em jogo. Foi aqui que criamos funções capazes de comprimir tanto pecas de um mapa como jogadores e os disparos. Sendo que para facilitar mais tarde a descompressão utilizamos caracteres de separação dados ao acaso para separar num estado cada componente os seus elementos. Para a descompressão fez-se exatamente o oposto da função anterior sendo que descomprime o estado anteriormente comprimido. Tal como a comprime, esta função encontra-se dividida por outras funções auxiliares que individualmente descomprimem ora pecas do mapa ora jogadores e disparos. Nesta função cada caracter anteriormente atribuído é descomprimido para o seu estado normal, sendo que os caracteres de separação usados na compressão são tidos em conta.

Conclusao:

No fim desta tarefa foi-nos possível comprimir um estado do jogo em pouco mais que alguns caracteres. Consequentemente, foi-nos possível também a descompressão desse mesmo estado, anteriormente comprimido, devolvendo-o ao seu estado inicial. Ou seja fomos capazes de, usando a funçao comprime transformar um Estado em nada mais que uma String, e usando a funçao descomprime transformar essa mesma 'String 'de novo num Estado igual ao fornecido inicialmente premitindo-nos assim poupar espaço no disco aquando de ser gravado.

Module : Tarefa5_2018li1g107

ID : Maria Sofia Martinho Gonçalves Jordao Marques (a87963); Jose Nuno Baptista Martins (a90122)

Data : 28.12.2018

Universidade do Minho

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3.

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> String

Comprime um Estado para formato textual.

NB: A função show representa um Estado num formato textual facilmente legível mas extenso.

NB: Uma boa solução deve representar o Estado dado no mínimo número de caracteres possível.

auxCMap :: [[Peca]] -> String

Dado um mapa, este separa as varias listas de Pecas com um caracter especifico que lhe fora atribuido.

auxM :: [Peca] -> String

Dada um lista de Pecas, esta funcao comprime numa determinada string cada uma das pecas ai presentes.

auxCJdsStr :: [Jogador] -> String

Uma lista de jogadores sera comprimida a uma string, em que cada character representara um dado Jogador. Nao esquecendo tambem que a separacao entre cada jogador da lista é tambem representado por um caracter especifico (primitindo mais tarde uma descompressao mais "fácil").

auxCDisp :: [Disparo] -> String

Esta auxiliar premite nos do mesmo modo, comprimir a uma string, uma dada lista de Disparos sendo que é tambem apresentada nessa string caracteres especificos representantes da separacao entre jogadores.

auxtd :: Direcao -> String

Esta funçao premite-nos nada mais do que comprimir as diferentes direçoes em strings proprias as mesmas.

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

NB: A função comprime é válida de for possível recuperar o Estado utilizando a função descomprime, i.e.:

descomprime . comprime = id

NB: Esta propriedade é particularmente válida para a solução pré-definida:

read . show = id

auxDMapa

Arguments

:: String
lista de Char (representa pecas do mapa)

-> Mapa
lista de lista de Pecas

-> Mapa
Mapa resultante da descompressao dessa string

É nesta funçao que ocorre a descompressao da string relativa ao mapa do jogo (resultante da anterior compressao do mapa ).

auxDescLinha :: String -> [Peca]

Esta funçao auxilia excluisavemente a funçao anterior sendo que o seu proposito é descomprimir a string em Pecas.

auxJo :: String -> Char -> Int

Auxiliar que premite ler uma determinada string ate um certo ponto de paragem, algo extremamente util para a funçao descomprimir em si.

auxDJg

Arguments

:: String
string resultante do 1º drop da string anterior

-> [Jogador]

-> [Jogador]

Lê a string correspondente à lista dos jogadores

auxJog :: String -> Jogador

Descomprime uma parte da string relativa a um só Jogador

isnumero :: Char -> Bool

Esta funçao verifica se um dado Char é ou nao um algarismo ( tendo em conta a numenclatura da Tabela ASCII)

auxENum :: String -> String -> Int

Lê ate onde é que a string é um algarismo, tendo como auxiliar a funçao definida anteriormente.

auxD :: String -> Direcao

Ao contrario da funçao auxtd, esta faz o oposto, descomprimindo cada string numa direcao correspondente.

auxDDisp :: String -> Disparo

Descomprime a string relativa a um disparo no próprio Disparo.

auxDF :: String -> [Disparo] -> [Disparo]

Descomprime a string direcionada para o lista dos disparos na propria lista de disparos que lhe deu origem.