Tarefa4_2018li1g107

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

map89 :: [[Peca]]

Mapa escolhido para constituir o Estado do teste.

jds :: [Jogador]

Lista de Jogadores ecolhida para constituir o Estado do teste.

disp :: [Disparo]

Lista de Disparos escolhida para constituir o Estado do teste

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

mapaatualizado :: [Disparo] -> Posicao -> [[Peca]] -> Mapa

É atraves desta funçao que o mapa do jogo vai sendo atualizado apos jogadas do tipo DisparoLaser

atualizajds :: [Disparo] -> [Jogador] -> [Jogador]

Esta funcao é utilizada com o objectivo de atualizar cada jogador, quanto ao seu numero de vidas, apos jogadas feitas do tipo DisparoLaser .

destroiTiroLaser :: [Disparo] -> [Disparo]

Funçao que simula a destruiçao de disparos quando se deparam com um disparo do tipo Laser.

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

mapaatualizado2 :: [Disparo] -> Posicao -> [[Peca]] -> Mapa

Tal como a funçao mapaatualizado, esta funçao premite-nos obter o mapa do jogo atualizado apos jogadas, neste caso, jogadas do tipo Disparo.

desapareceTiro :: [Disparo] -> Posicao -> [[Peca]] -> [Disparo]

Funçao que provoca o desaparecimento de um tiro do tipo DisparoCanhao quando colide com 'Blocos Destrutiveis' ou 'Blocos Indestrutiveis'.

atualizajds2 :: [Disparo] -> [Jogador] -> [Jogador]

Funçao auxiliar semelhante a atualizajds, que atualiza o numero de vida de cada jogador, apos jogadas do tipo DisparoCanhao.

desapareceTiro2 :: [Disparo] -> [Jogador] -> [Disparo]

Funçao que provoca o desaparecimento de um tiro do tipo DisparoCanhao quando atinge um Jogador.

destroiTiroCanhao :: [Disparo] -> [Disparo]

Funçao que simula a destruiçao de disparos quando se deparam com um disparo do tipo Canhao.

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

auxTicks :: [Disparo] -> [Disparo]

Auxiliar da funçao ticksChoques que decrementa o tempo do choque conforme o passar do tempo do jogo.

destroiTiroChoques :: [Disparo] -> [Disparo]

Tal como as funcoes 'dispara tiro' anteriores estas incrementa a situaçao onde um outro tiro entra na area abrangente onde um tiro choque esta a decorrer.