Tarefa6_2018li1g107

Description

Este módulo define funções comuns da Tarefa 6 do trabalho prático.

Introduçao:

O objectivo desta tarefa é implementar um robô capaz de jogar automaticamente tendo em base uma certa estratégia.

Objetivo:

Procurando tornar o nosso bot o mais autónomo que conseguimos, começamos por definir uma função capaz de avaliar a melhor jogada a fazer tendo em conta as outras posições dos jogadores. O mesmo fizemos para ganhar pontos através da destruição de blocos destrutíveis, ou seja, através de uma outra função conseguimos fazer com que um bot de acordo com a sua posição e o mapa em questão, tomasse a decisão de contornar blocos indestrutíveis, se perante os mesmos, e destruir os restantes.

Conclusao:

Nesta tarefa ficamos aquem daquilo que pretendiamos, que seria a autonomizaçao do nosso bot No entanto, apesar das suas falhas, concluimos o trabalho que consistiu em criar um jogo 2D de combate.

Module : Tarefa5_2018li1g107

ID : Maria Sofia Martinho Gonçalves Jordao Marques (a87963); Jose Nuno Baptista Martins (a90122)

Data : 28.12.2018

Universidade do Minho

data Deiricao

Constructors

Acima

Abaixo

Aiesquerda

Aidireita

Aladonenhum

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

map89 :: [[Peca]]

jds :: [Jogador]

disp :: [Disparo]

atingeJogadores :: Int -> [Jogador] -> Mapa -> Maybe Jogada

Avalia,com ajuda da funçao auxiliar jogadoresEmLinha, a melhor Jogada do tipo 'Dispara Arma' a fazer , por um determinado bot, tendo em conta as posicoes dos outros jogadores e a situacao em que o proprio bot assim se encontra.

jogadoresEmLinha :: Int -> Direcao -> [Jogador] -> Deiricao

Funcao que retorna se existem tanks numa direcao escolhida.

defendeAtaque :: Jogador -> [Disparo] -> Maybe Jogada

desviaBlocos :: Int -> [Jogador] -> Mapa -> Maybe Jogada

Funçao onde o bot de acordo com a sua posição e o mapa em questão, toma a decisão de contornar blocos indestrutíveis, se perante os mesmos, e destruir os blocos Destrutiveis.